Initial support for JSON help messages ##json
