+++ libr/cons/Makefile
@@ -3,7 +3,7 @@ include ../config.mk
 NAME=r_cons
 OBJS=cons.o cpipe.o output.o grep.o less.o more.o pager.o cutf8.o
 OBJS+=line.o hud.o rgb.o input.o pal.o editor.o 2048.o html.o
-OBJS+=canvas.o canvas_line.o stiv.o pixel.o
+OBJS+=canvas.o canvas_line.o stiv.o pixel.o help.o
 R2DEPS=r_util
 
 include ../rules.mk
